requirements.txt
gunicorn==20.0.4
Flask==2.3.3
pandas==2.1.3
Flask-Cors==4.0.0
gensim==4.3.2
mecab-python3==1.0.8
scipy==1.11.4
transformers==4.35.2
torch2.1.1
